code-insight/query/requirements.txt
fastapi
uvicorn
neo4j
sentence-transformers
python-dotenv